Fix scpeech screen

+++ Rpg.Client/Rpg.Client/GameScreens/Speech/SpeechScreen.cs
@@ -373,7 +373,7 @@ namespace Rpg.Client.GameScreens.Speech
                 var textFragmentSize = textFragmentControl.CalculateSize();
 
                 const int SPEECH_MARGIN = 50;
-                var sumOptionHeight = _optionButtons.Sum(x => CalcOptionButtonSize(x).Y) + OPTION_BUTTON_MARGIN;
+                var sumOptionHeight = (int)_optionButtons.Sum(x => CalcOptionButtonSize(x).Y) + OPTION_BUTTON_MARGIN;
                 textFragmentControl.Rect = new Rectangle(
                     new Point(PORTRAIT_SIZE,
                         contentRectangle.Bottom - (int)textFragmentSize.Y - SPEECH_MARGIN - sumOptionHeight),
